code/requirements.txt
discord.py
aiohttp
websockets<7.0,>=6.0
pytz
tinydb